Refactor project layout
