--- include/redstone/net/network.hpp
@@ -1,7 +1,0 @@
-#pragma once
-
-namespace redstone::net {
-class network {
-public:
-};
-} // namespace redstone::net
