--- include/redstone/replica/machine.hpp
@@ -1,5 +1,0 @@
-#pragma once
-
-namespace redstone::replica {
-struct machine {};
-} // namespace redstone::replica
