+++ src/random/splitmix.hpp
