+++ src/sim/file_descriptor.cpp
@@ -1,4 +1,4 @@
-#include "redstone/sim/file_descriptor.hpp"
+#include "sim/file_descriptor.hpp"
 
 #include <cassert>
